requirements.txt
pyserial
pytest
typer[all]
dependency-injector
